Minor GUI changes
