+++ frontend/package.json
@@ -25,6 +25,7 @@
     "react-rating": "^2.0.5",
     "react-router-dom": "^6.2.1",
     "react-scripts": "5.0.1",
+    "react-star-ratings": "^2.3.0",
     "reactjs-popup": "^2.0.6",
     "web-vitals": "^2.1.4"
   },
